Fixed SIP formula, ui changes, etc

+++ app/src/main/java/com/example/sipcalculator/SipModel.kt
@@ -2,6 +2,6 @@ package com.example.sipcalculator
 
 data class SipModel(
     val years: Int,
-    val invested: Int,
-    val finalAmount: Int
+    val invested: Double,
+    val finalAmount: Double
 )
